setup ingress / setup project deployment and service / setup skaffold for local work / update env variables usage / set test env variables / setup dockerfile / setup permission-check middleware

+++ src/controllers/email.controller.ts
@@ -1,35 +1,28 @@
 import { Request, Response } from 'express';
+import { isValidObjectId } from 'mongoose';
 import { buildAttachments } from '../helpers/build-attachments';
 import { dataResponseFormat } from '../helpers/data-response-format';
 import { Email } from '../models/email.model';
 
-export const getEmails = async (req: Request, res: Response) => {
-  const emails = await Email.find();
-  const emailsFormattedData = await dataResponseFormat(req, emails);
-  res.send(emailsFormattedData);
-};
-
-export const getEmailById = async (req: Request, res: Response) => {
-  const emailId = req.params.id;
-  const email = await Email.findById(emailId);
-
-  if (!email) {
-    return res.status(404).send('Not found');
-  }
-
-  res.send(email);
-};
-
 export const createEmail = async (req: Request, res: Response) => {
   const attachments: string[] = buildAttachments(req);
-  const email = Email.build({ ...req.body, userId: '3123133', attachments });
+  const email = Email.build({
+    ...req.body,
+    userId: req.currentUser.id,
+    attachments,
+  });
   await email.save();
 
-  res.status(201).send({ status: 'OK' });
+  res.status(201).send(email);
 };
 
 export const updateEmail = async (req: Request, res: Response) => {
   const emailId = req.params.id;
+
+  if (!isValidObjectId(emailId)) {
+    return res.status(404).send('Not found, Invalid id provided');
+  }
+
   const email = await Email.findById(emailId);
 
   if (!email) {
@@ -44,6 +37,11 @@ export const updateEmail = async (req: Request, res: Response) => {
 
 export const deleteEmail = async (req: Request, res: Response) => {
   const emailId = req.params.id;
+
+  if (!isValidObjectId(emailId)) {
+    return res.status(404).send('Not found, Invalid id provided');
+  }
+
   const email = await Email.findByIdAndRemove(emailId);
 
   if (!email) {
@@ -51,4 +49,26 @@ export const deleteEmail = async (req: Request, res: Response) => {
   }
 
   res.sendStatus(204);
+};
+
+export const getEmails = async (req: Request, res: Response) => {
+  const emails = await Email.find();
+  const emailsFormattedData = await dataResponseFormat(req, emails);
+  res.send(emailsFormattedData);
+};
+
+export const getEmailById = async (req: Request, res: Response) => {
+  const emailId = req.params.id;
+
+  if (!isValidObjectId(emailId)) {
+    return res.status(404).send('Not found, Invalid id provided');
+  }
+
+  const email = await Email.findById(emailId);
+
+  if (!email) {
+    return res.status(404).send('Not found');
+  }
+
+  res.send(email);
 };
